requirements.txt
# Core Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# AI/ML
google-generativeai>=0.3.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0
python-dateutil>=2.8.0

# Database
SQLAlchemy==2.0.46

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Web Framework
streamlit>=1.28.0
